Separate files of getSpotGroupServers API.
