+++ src/routes/spot-automation/spot-group.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 import * as spotGroup from '@controllers/spot-automation/spot-group';
+import getSpotGroupServers from '@controllers/spot-automation/spot-group/get-spot-group-servers';
 import listSpotGroupMembers from '@controllers/spot-automation/spot-group/list-spot-group-members';
 
 const router = express.Router();
@@ -15,7 +16,7 @@ const controllers = [
     { url: '/stat', func: spotGroup.statSpotGroups },
     { url: '/get-candidates', func: spotGroup.getCandidates },
     { url: '/get-supported-resource-types', func: spotGroup.getSupportedResourceTypes },
-    { url: '/get-spot-group-servers', func: spotGroup.getSpotGroupServers },
+    { url: '/get-spot-group-servers', func: getSpotGroupServers },
     { url: '/member/list', func: listSpotGroupMembers}
 ];
